user_management_service/requirements.txt
Flask==2.1.1
Werkzeug==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.3.1
python-dotenv==0.19.2
psycopg2-binary==2.9.3
SQLAlchemy==1.4.46
requests==2.26.0